Split definitions into a header file.
